requirements.txt
requests
openai
PyPDF2
python-dotenv
